moved all radiation functions from SpinTemperatureBox.c to XraySourceBox.c
